% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pseudo_bulk_samples.R
\name{.default_cells_per_sample}
\alias{.default_cells_per_sample}
\title{Default number of cells per pseudo-bulk sample}
\usage{
.default_cells_per_sample(meta, barcode_col)
}
\arguments{
\item{meta}{Cell metadata \code{data.frame}.}

\item{barcode_col}{Column name for barcode identifiers.}
}
\value{
Integer scalar.
}
\description{
Returns the median barcode cell count in \code{meta}.
}
\keyword{internal}
