% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT.R
\name{.enable_parallel_cibersort}
\alias{.enable_parallel_cibersort}
\title{Configure a furrr parallel backend for CIBERSORT}
\usage{
.enable_parallel_cibersort(nThreads = NULL, maxSize = 500)
}
\arguments{
\item{nThreads}{Number of parallel workers. Defaults to \code{availableCores() - 2}.}

\item{maxSize}{Maximum global object size (MB) for future exports.}
}
\description{
Configure a furrr parallel backend for CIBERSORT
}
\keyword{internal}
