% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_scorings.R
\name{eval_MAE}
\alias{eval_MAE}
\title{Mean absolute error for cellular proportion vectors}
\usage{
eval_MAE(p_obs, p_estimated, trim_shared_zeros = TRUE)
}
\value{
A numeric scalar score.
}
\description{
Mean absolute error for cellular proportion vectors
}
