% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_scorings.R
\name{eval_RMSE}
\alias{eval_RMSE}
\title{Root mean squared error for cellular proportion vectors}
\usage{
eval_RMSE(p_obs, p_estimated, trim_shared_zeros = TRUE)
}
\value{
A numeric scalar score.
}
\description{
Root mean squared error for cellular proportion vectors
}
