% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GastroDeconv2FateMap-package.R
\docType{package}
\name{GastroDeconv2FateMap-package}
\alias{GastroDeconv2FateMap}
\alias{GastroDeconv2FateMap-package}
\title{GastroDeconv2FateMap: Pseudo-Bulk Simulation and Deconvolution Benchmarking for Gastruloids}
\description{
Utilities for aggregating single-cell profiles into pseudo-bulk samples, annotating mouse genes, scoring deconvolution performance, and benchmarking phenotype-guided algorithms on gastruloid differentiation data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bastienchassagnol/GastroDeconv2FateMap}
  \item \url{https://bastienchassagnol.github.io/GastroDeconv2FateMap/}
  \item Report bugs at \url{https://github.com/bastienchassagnol/GastroDeconv2FateMap/issues}
}

}
\author{
\strong{Maintainer}: Bastien Chassagnol \email{chassagnolbastien@gmail.com}

Authors:
\itemize{
  \item Bastien Chassagnol \email{chassagnolbastien@gmail.com}
}

}
\keyword{internal}
