% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_features_pub}
\alias{theme_features_pub}
\title{Publication-sized variant of \code{theme_features()}}
\usage{
theme_features_pub(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{theme_features()}.}
}
\value{
A \strong{ggplot2} \code{theme} object.
}
\description{
Publication-sized variant of \code{theme_features()}
}
