% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_scorings.R
\name{eval_SDID}
\alias{eval_SDID}
\title{Square-root sine distance between two compositional vectors}
\usage{
eval_SDID(p_obs, p_estimated, trim_shared_zeros = TRUE)
}
\value{
A numeric scalar score.
}
\description{
Square-root sine distance between two compositional vectors
}
\references{
\url{https://mpra.ub.uni-muenchen.de/84387/}
}
